// apps/core-service/src/app/modules/authentication/modules/mfa/dto/Device.dto.ts
//=============================================================================
//Licensed Materials - Property of Bambee
//(C) Copyright Bambee 2022
//All Rights Reserved
//=============================================================================

export interface DeviceDTO {
  readonly id: string;
  readonly userAgent: string;
}