// apps/core-service/src/app/modules/authentication/modules/mfa/types/MfaType.ts
//=============================================================================
//Licensed Materials - Property of Bambee
//(C) Copyright Bambee 2022
//All Rights Reserved
//=============================================================================

export enum MfaType {
  SMS = 'sms',
  BACKUP = 'backup',
  PASSWORD = 'password',
}